install dependencies on circle

+++ spec/services/csv_to_s3_exporter_spec.rb
@@ -4,8 +4,8 @@ describe CsvToS3Exporter do
   describe "#call", skip_db_cleaner: true do # skip cleaner so our data is available to psql CLI
     after do
       DatabaseCleaner.clean
-      File.unlink(temp_file)
-      File.unlink(meta_file)
+      File.unlink(temp_file) if File.exist?(temp_file)
+      File.unlink(meta_file) if File.exist?(meta_file)
       fail "failed to clean up #{temp_file}" if File.exist?(temp_file)
       fail "failed to clean up #{meta_file}" if File.exist?(meta_file)
     end
